fix: re-add mark.js to users list for display name
